# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Admin Backend API",
        "description": "RESTful API for managing portfolio admin dashboard content with Clean Architecture implementation",
        "contact": {
            "name": "Nhật Anh Dev",
            "email": "admin@nhatanh.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/hero": {
            "get": {
                "tags": [
                    "Hero Section"
                ],
                "summary": "Get hero section content",
                "description": "Retrieve current hero section content in both Vietnamese and English",
                "operationId": "290db80790dd1bb62da88d70c1626fd3",
                "responses": {
                    "200": {
                        "description": "Hero content retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "greeting_vi": {
                                                    "type": "string",
                                                    "example": "Xin chào, tôi là"
                                                },
                                                "greeting_en": {
                                                    "type": "string",
                                                    "example": "Hello, I'm"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nhật Anh"
                                                },
                                                "title_vi": {
                                                    "type": "string",
                                                    "example": "Lập trình viên Fullstack"
                                                },
                                                "title_en": {
                                                    "type": "string",
                                                    "example": "Fullstack Developer"
                                                },
                                                "subtitle_vi": {
                                                    "type": "string",
                                                    "example": "Chuyên về phát triển web hiện đại"
                                                },
                                                "subtitle_en": {
                                                    "type": "string",
                                                    "example": "Specialized in modern web development"
                                                },
                                                "cta_text_vi": {
                                                    "type": "string",
                                                    "example": "Xem dự án"
                                                },
                                                "cta_text_en": {
                                                    "type": "string",
                                                    "example": "View Projects"
                                                },
                                                "cta_link": {
                                                    "type": "string",
                                                    "example": "#projects"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hero Section"
                ],
                "summary": "Update hero section content",
                "description": "Update hero section content with bilingual support",
                "operationId": "fa519bb3e4f39959ca7e515f4ff243a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "greeting_vi",
                                    "greeting_en",
                                    "name",
                                    "title_vi",
                                    "title_en",
                                    "subtitle_vi",
                                    "subtitle_en",
                                    "cta_text_vi",
                                    "cta_text_en",
                                    "cta_link"
                                ],
                                "properties": {
                                    "greeting_vi": {
                                        "type": "string",
                                        "example": "Xin chào, tôi là"
                                    },
                                    "greeting_en": {
                                        "type": "string",
                                        "example": "Hello, I'm"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nhật Anh"
                                    },
                                    "title_vi": {
                                        "type": "string",
                                        "example": "Lập trình viên Fullstack"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "example": "Fullstack Developer"
                                    },
                                    "subtitle_vi": {
                                        "type": "string",
                                        "example": "Chuyên về phát triển web hiện đại"
                                    },
                                    "subtitle_en": {
                                        "type": "string",
                                        "example": "Specialized in modern web development"
                                    },
                                    "cta_text_vi": {
                                        "type": "string",
                                        "example": "Xem dự án"
                                    },
                                    "cta_text_en": {
                                        "type": "string",
                                        "example": "View Projects"
                                    },
                                    "cta_link": {
                                        "type": "string",
                                        "example": "#projects"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hero content updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hero content updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "greeting_vi": {
                                                    "type": "string",
                                                    "example": "Xin chào, tôi là"
                                                },
                                                "greeting_en": {
                                                    "type": "string",
                                                    "example": "Hello, I'm"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nhật Anh"
                                                },
                                                "title_vi": {
                                                    "type": "string",
                                                    "example": "Lập trình viên Fullstack"
                                                },
                                                "title_en": {
                                                    "type": "string",
                                                    "example": "Fullstack Developer"
                                                },
                                                "subtitle_vi": {
                                                    "type": "string",
                                                    "example": "Chuyên về phát triển web hiện đại"
                                                },
                                                "subtitle_en": {
                                                    "type": "string",
                                                    "example": "Specialized in modern web development"
                                                },
                                                "cta_text_vi": {
                                                    "type": "string",
                                                    "example": "Xem dự án"
                                                },
                                                "cta_text_en": {
                                                    "type": "string",
                                                    "example": "View Projects"
                                                },
                                                "cta_link": {
                                                    "type": "string",
                                                    "example": "#projects"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}